Split #7 + #8: extract engine/lsm/iterator and engine/lsm/flush
iterator subpackage owns the generic merge/concat machinery:

  engine/lsm/iterator/
    concat.go      ConcatIterator over []*table.Table
    concat_test.go
    merge.go       MergeIterator + emptyIterator + node fan-in
    merge_test.go

The lsm package keeps LSM.NewIterators and memTable.NewIterator since
they reach into shard internals. local/internal/iterator switches to
the new path via a lsmiter alias (avoids name clash with its own
package iterator).

flush subpackage owns the per-shard flush queue with the WAL-segment
ordering invariant. The Runtime is decoupled from *memTable: callers
pass a shardID and an opaque Payload, which the lsm worker loop
asserts back to *memTable. Errors are renamed to flush.Err{Nil,Closed,
NilPayload}; local/errkind/classify follows.

  engine/lsm/flush/
    runtime.go      Runtime + Task + per-shard FIFO + Stats
    runtime_test.go enqueue/serialization/parallelism invariants
