+++ engine/lsm/iterator.go
@@ -1,27 +1,8 @@
 package lsm
 
 import (
-	"bytes"
-	"fmt"
-
 	"github.com/feichai0017/NoKV/engine/index"
-	"github.com/feichai0017/NoKV/engine/kv"
-	"github.com/feichai0017/NoKV/engine/lsm/table"
-	"github.com/feichai0017/NoKV/utils"
 )
-
-type emptyIterator struct{}
-
-type entryItem struct{ e *kv.Entry }
-
-func (it *entryItem) Entry() *kv.Entry { return it.e }
-
-func (*emptyIterator) Next()            {}
-func (*emptyIterator) Rewind()          {}
-func (*emptyIterator) Seek([]byte)      {}
-func (*emptyIterator) Valid() bool      { return false }
-func (*emptyIterator) Item() index.Item { return nil }
-func (*emptyIterator) Close() error     { return nil }
 
 // NewIterators builds iterators over mutable/immutable memtables across all
 // shards plus the SST levels. Per-shard ordering is preserved (active first,
@@ -48,7 +29,6 @@ func (lsm *LSM) NewIterators(opt *index.Options) []index.Iterator {
 	return iters
 }
 
-// memtable iterator
 type memIterator struct {
 	innerIter index.Iterator
 }
@@ -65,7 +45,6 @@ func (m *memTable) NewIterator(opt *index.Options) index.Iterator {
 	return &memIterator{innerIter: inner}
 }
 
-// Next advances the memtable iterator.
 func (iter *memIterator) Next() {
 	if iter.innerIter == nil {
 		return
@@ -73,7 +52,6 @@ func (iter *memIterator) Next() {
 	iter.innerIter.Next()
 }
 
-// Valid reports whether the memtable iterator is valid.
 func (iter *memIterator) Valid() bool {
 	if iter.innerIter == nil {
 		return false
@@ -81,7 +59,6 @@ func (iter *memIterator) Valid() bool {
 	return iter.innerIter.Valid()
 }
 
-// Rewind resets the memtable iterator to the beginning.
 func (iter *memIterator) Rewind() {
 	if iter.innerIter == nil {
 		return
@@ -89,7 +66,6 @@ func (iter *memIterator) Rewind() {
 	iter.innerIter.Rewind()
 }
 
-// Item returns the current memtable item.
 func (iter *memIterator) Item() index.Item {
 	if iter.innerIter == nil {
 		return nil
@@ -97,7 +73,6 @@ func (iter *memIterator) Item() index.Item {
 	return iter.innerIter.Item()
 }
 
-// Close releases resources held by the underlying index iterator.
 func (iter *memIterator) Close() error {
 	if iter.innerIter == nil {
 		return nil
@@ -105,447 +80,9 @@ func (iter *memIterator) Close() error {
 	return iter.innerIter.Close()
 }
 
-// Seek positions the memtable iterator at key.
 func (iter *memIterator) Seek(key []byte) {
 	if iter.innerIter == nil {
 		return
 	}
 	iter.innerIter.Seek(key)
-}
-
-// ConcatIterator merge multiple iterators into one
-type ConcatIterator struct {
-	idx     int // Which iterator is active now.
-	cur     index.Iterator
-	tables  []*table.Table       // Disregarding reversed, this is in ascending order.
-	options *index.Options // Valid options are REVERSED and NOCACHE.
-}
-
-// NewConcatIterator creates a new concatenated iterator
-func NewConcatIterator(tbls []*table.Table, opt *index.Options) *ConcatIterator {
-	return &ConcatIterator{
-		options: opt,
-		tables:  tbls,
-		idx:     -1, // Not really necessary because s.it.Valid()=false, but good to have.
-	}
-}
-
-func (s *ConcatIterator) setIdx(idx int) {
-	if idx == s.idx && s.cur != nil {
-		return
-	}
-	_ = s.closeCurrent()
-	s.idx = idx
-	if idx < 0 || idx >= len(s.tables) {
-		s.cur = nil
-		return
-	}
-	s.cur = s.tables[idx].NewIterator(s.options)
-}
-
-func (s *ConcatIterator) closeCurrent() error {
-	if s.cur == nil {
-		return nil
-	}
-	err := s.cur.Close()
-	s.cur = nil
-	return err
-}
-
-// Rewind implements Interface
-func (s *ConcatIterator) Rewind() {
-	if len(s.tables) == 0 {
-		return
-	}
-	if s.options.IsAsc {
-		s.setIdx(0)
-	} else {
-		s.setIdx(len(s.tables) - 1)
-	}
-	if s.cur == nil {
-		return
-	}
-	s.cur.Rewind()
-	s.advanceToValidCurrent()
-}
-
-// Valid implements y.Interface
-func (s *ConcatIterator) Valid() bool {
-	return s.cur != nil && s.cur.Valid()
-}
-
-// Item _
-func (s *ConcatIterator) Item() index.Item {
-	return s.cur.Item()
-}
-
-// Seek brings us to element >= key if reversed is false. Otherwise, <= key.
-func (s *ConcatIterator) Seek(key []byte) {
-	n := len(s.tables)
-	if n == 0 {
-		s.setIdx(-1)
-		return
-	}
-	var idx int
-	if s.options.IsAsc {
-		// First table whose max key >= target.
-		lo, hi := 0, n
-		for lo < hi {
-			mid := lo + (hi-lo)/2
-			if kv.CompareInternalKeys(s.tables[mid].MaxKey(), key) >= 0 {
-				hi = mid
-			} else {
-				lo = mid + 1
-			}
-		}
-		idx = lo
-	} else {
-		// Last table whose min key <= target.
-		lo, hi := 0, n
-		for lo < hi {
-			mid := lo + (hi-lo)/2
-			if kv.CompareInternalKeys(s.tables[mid].MinKey(), key) > 0 {
-				hi = mid
-			} else {
-				lo = mid + 1
-			}
-		}
-		idx = lo - 1
-	}
-	if idx >= len(s.tables) || idx < 0 {
-		s.setIdx(-1)
-		return
-	}
-	// For reversed=false, we know s.tables[i-1].Biggest() < key. Thus, the
-	// previous table cannot possibly contain key.
-	s.setIdx(idx)
-	if s.cur == nil {
-		return
-	}
-	s.cur.Seek(key)
-	s.advanceToValidCurrent()
-}
-
-// Next advances our concat iterator.
-func (s *ConcatIterator) Next() {
-	s.cur.Next()
-	if s.cur.Valid() {
-		// Nothing to do. Just stay with the current table.
-		return
-	}
-	for { // In case there are empty tables.
-		if s.options.IsAsc {
-			s.setIdx(s.idx + 1)
-		} else {
-			s.setIdx(s.idx - 1)
-		}
-		if s.cur == nil {
-			// End of list. Valid will become false.
-			return
-		}
-		s.cur.Rewind()
-		if s.cur.Valid() {
-			break
-		}
-	}
-}
-
-func (s *ConcatIterator) advanceToValidCurrent() {
-	for s.cur != nil && !s.cur.Valid() {
-		if s.options.IsAsc {
-			s.setIdx(s.idx + 1)
-		} else {
-			s.setIdx(s.idx - 1)
-		}
-		if s.cur == nil {
-			return
-		}
-		s.cur.Rewind()
-	}
-}
-
-// Close implements y.Interface.
-func (s *ConcatIterator) Close() error {
-	if err := s.closeCurrent(); err != nil {
-		return fmt.Errorf("ConcatIterator:%+v", err)
-	}
-	return nil
-}
-
-// MergeIterator merges multiple iterators into a single ordered stream.
-// NOTE: MergeIterator owns the array of iterators and is responsible for closing them.
-type MergeIterator struct {
-	left  node
-	right node
-	small *node
-
-	curKey  []byte
-	reverse bool
-}
-
-type node struct {
-	valid bool
-	entry *kv.Entry
-	iter  index.Iterator
-
-	// The two iterators are type asserted from `y.Iterator`, used to inline more function calls.
-	// Calling functions on concrete types is much faster (about 25-30%) than calling the
-	// interface's function.
-	merge  *MergeIterator
-	concat *ConcatIterator
-}
-
-func (n *node) setFromMerge() {
-	n.valid = n.merge.small != nil && n.merge.small.valid && n.merge.small.entry != nil
-	n.entry = nil
-	if n.valid {
-		n.entry = n.merge.small.entry
-	}
-}
-
-func (n *node) setFromConcat() {
-	n.valid = n.concat.Valid()
-	n.entry = nil
-	if !n.valid {
-		return
-	}
-	item := n.concat.Item()
-	utils.CondPanicFunc(item == nil, func() error {
-		return fmt.Errorf("concat iterator valid but item nil (type=%T)", n.concat)
-	})
-	entry := item.Entry()
-	utils.CondPanicFunc(entry == nil, func() error {
-		return fmt.Errorf("concat iterator valid but entry nil (type=%T)", n.concat)
-	})
-	n.entry = entry
-}
-
-func (n *node) setFromIter() {
-	n.valid = n.iter.Valid()
-	n.entry = nil
-	if !n.valid {
-		return
-	}
-	item := n.iter.Item()
-	utils.CondPanicFunc(item == nil, func() error {
-		return fmt.Errorf("iterator valid but item nil (type=%T)", n.iter)
-	})
-	entry := item.Entry()
-	utils.CondPanicFunc(entry == nil, func() error {
-		return fmt.Errorf("iterator valid but entry nil (type=%T)", n.iter)
-	})
-	n.entry = entry
-}
-
-func (n *node) setIterator(iter index.Iterator) {
-	n.iter = iter
-	// It's okay if the type assertion below fails and n.merge/n.concat are set to nil.
-	// We handle the nil values of merge and concat in all the methods.
-	n.merge, _ = iter.(*MergeIterator)
-	n.concat, _ = iter.(*ConcatIterator)
-}
-
-func (n *node) setKey() {
-	if n.iter == nil {
-		n.valid = false
-		n.entry = nil
-		return
-	}
-	switch {
-	case n.merge != nil:
-		n.setFromMerge()
-	case n.concat != nil:
-		n.setFromConcat()
-	default:
-		n.setFromIter()
-	}
-}
-
-func (n *node) next() {
-	if n.iter == nil {
-		n.valid = false
-		n.entry = nil
-		return
-	}
-	switch {
-	case n.merge != nil:
-		n.merge.Next()
-		n.setFromMerge()
-	case n.concat != nil:
-		n.concat.Next()
-		n.setFromConcat()
-	default:
-		n.iter.Next()
-		n.setFromIter()
-	}
-}
-
-func (n *node) rewind() {
-	if n.iter == nil {
-		n.valid = false
-		n.entry = nil
-		return
-	}
-	n.iter.Rewind()
-	n.setKey()
-}
-
-func (n *node) seek(key []byte) {
-	if n.iter == nil {
-		n.valid = false
-		n.entry = nil
-		return
-	}
-	n.iter.Seek(key)
-	n.setKey()
-}
-
-func (mi *MergeIterator) fix() {
-	if !mi.bigger().valid {
-		return
-	}
-	if !mi.small.valid {
-		mi.swapSmall()
-		return
-	}
-	cmp := kv.CompareInternalKeys(mi.small.entry.Key, mi.bigger().entry.Key)
-	switch {
-	case cmp == 0: // Both the keys are equal.
-		// In case of same keys, move the right iterator ahead.
-		mi.right.next()
-		if &mi.right == mi.small {
-			mi.swapSmall()
-		}
-		return
-	case cmp < 0: // Small is less than bigger().
-		if mi.reverse {
-			mi.swapSmall()
-		}
-		return
-	default: // bigger() is less than small.
-		if mi.reverse {
-			// Do nothing since we're iterating in reverse. Small currently points to
-			// the bigger key and that's okay in reverse iteration.
-		} else {
-			mi.swapSmall()
-		}
-		return
-	}
-}
-
-func (mi *MergeIterator) bigger() *node {
-	if mi.small == &mi.left {
-		return &mi.right
-	}
-	return &mi.left
-}
-
-func (mi *MergeIterator) swapSmall() {
-	if mi.small == &mi.left {
-		mi.small = &mi.right
-		return
-	}
-	if mi.small == &mi.right {
-		mi.small = &mi.left
-		return
-	}
-}
-
-// Next returns the next element. If it is the same as the current key, ignore it.
-func (mi *MergeIterator) Next() {
-	for mi.Valid() {
-		if !bytes.Equal(mi.small.entry.Key, mi.curKey) {
-			break
-		}
-		mi.small.next()
-		mi.fix()
-	}
-	mi.setCurrent()
-}
-
-func (mi *MergeIterator) setCurrent() {
-	if mi.small.valid && mi.small.entry == nil {
-		mi.small.valid = false
-		mi.small.entry = nil
-		return
-	}
-	if mi.small.valid {
-		mi.curKey = append(mi.curKey[:0], mi.small.entry.Key...)
-	}
-}
-
-// Rewind seeks to first element (or last element for reverse iterator).
-func (mi *MergeIterator) Rewind() {
-	mi.left.rewind()
-	mi.right.rewind()
-	mi.fix()
-	mi.setCurrent()
-}
-
-// Seek brings us to element with key >= given key.
-func (mi *MergeIterator) Seek(key []byte) {
-	mi.left.seek(key)
-	mi.right.seek(key)
-	mi.fix()
-	mi.setCurrent()
-}
-
-// Valid returns whether the MergeIterator is at a valid element.
-func (mi *MergeIterator) Valid() bool {
-	return mi.small.valid
-}
-
-// Key returns the key associated with the current iterator.
-func (mi *MergeIterator) Item() index.Item {
-	return mi.small.iter.Item()
-}
-
-// Close implements Iterator.
-func (mi *MergeIterator) Close() error {
-	err1 := mi.left.iter.Close()
-	err2 := mi.right.iter.Close()
-	if err1 != nil {
-		return fmt.Errorf("merge iterator close left: %w", err1)
-	}
-	if err2 != nil {
-		return fmt.Errorf("merge iterator close right: %w", err2)
-	}
-	return nil
-}
-
-// NewMergeIterator creates a merge iterator.
-func NewMergeIterator(iters []index.Iterator, reverse bool) index.Iterator {
-	filtered := iters[:0]
-	for _, it := range iters {
-		if it != nil {
-			filtered = append(filtered, it)
-		}
-	}
-	return newMergeIterator(filtered, reverse)
-}
-
-func newMergeIterator(iters []index.Iterator, reverse bool) index.Iterator {
-	switch len(iters) {
-	case 0:
-		return &emptyIterator{}
-	case 1:
-		return iters[0]
-	case 2:
-		mi := &MergeIterator{
-			reverse: reverse,
-		}
-		mi.left.setIterator(iters[0])
-		mi.right.setIterator(iters[1])
-		// Assign left iterator randomly. This will be fixed when user calls rewind/seek.
-		mi.small = &mi.left
-		return mi
-	}
-	mid := len(iters) / 2
-	mi := &MergeIterator{
-		reverse: reverse,
-	}
-	mi.left.setIterator(newMergeIterator(iters[:mid], reverse))
-	mi.right.setIterator(newMergeIterator(iters[mid:], reverse))
-	mi.small = &mi.left
-	return mi
 }
