+++ engine/lsm/iterator_test.go
@@ -2,7 +2,6 @@ package lsm
 
 import (
 	"bytes"
-	"sort"
 	"testing"
 
 	"github.com/feichai0017/NoKV/engine/index"
@@ -12,136 +11,11 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-type sliceItem struct {
-	entry *kv.Entry
-}
-
-func (si sliceItem) Entry() *kv.Entry {
-	return si.entry
-}
-
-type sliceIterator struct {
-	entries []*kv.Entry
-	idx     int
-	reverse bool
-}
-
 func splitIterUserKey(t *testing.T, internal []byte) []byte {
 	t.Helper()
 	_, userKey, _, ok := kv.SplitInternalKey(internal)
 	require.True(t, ok)
 	return userKey
-}
-
-func (it *sliceIterator) Next() {
-	if it.reverse {
-		it.idx--
-		return
-	}
-	if it.idx < len(it.entries) {
-		it.idx++
-	}
-}
-
-func (it *sliceIterator) Valid() bool {
-	if it.reverse {
-		return it.idx >= 0 && it.idx < len(it.entries)
-	}
-	return it.idx < len(it.entries)
-}
-
-func (it *sliceIterator) Rewind() {
-	if it.reverse {
-		it.idx = len(it.entries) - 1
-		return
-	}
-	it.idx = 0
-}
-
-func (it *sliceIterator) Item() index.Item {
-	if !it.Valid() {
-		return nil
-	}
-	return sliceItem{entry: it.entries[it.idx]}
-}
-
-func (it *sliceIterator) Close() error {
-	return nil
-}
-
-func (it *sliceIterator) Seek(key []byte) {
-	if it.reverse {
-		it.idx = -1
-		for i := len(it.entries) - 1; i >= 0; i-- {
-			if bytes.Compare(it.entries[i].Key, key) <= 0 {
-				it.idx = i
-				break
-			}
-		}
-		return
-	}
-	it.idx = sort.Search(len(it.entries), func(i int) bool {
-		return bytes.Compare(it.entries[i].Key, key) >= 0
-	})
-}
-
-func TestMergeIteratorForwardAndReverse(t *testing.T) {
-	left := &sliceIterator{entries: []*kv.Entry{
-		{Key: kv.InternalKey(kv.CFDefault, []byte("a"), 1)},
-		{Key: kv.InternalKey(kv.CFDefault, []byte("c"), 1)},
-	}}
-	right := &sliceIterator{entries: []*kv.Entry{
-		{Key: kv.InternalKey(kv.CFDefault, []byte("b"), 1)},
-		{Key: kv.InternalKey(kv.CFDefault, []byte("c"), 1)},
-	}}
-
-	mi := NewMergeIterator([]index.Iterator{left, right}, false)
-	var keys []string
-	mi.Rewind()
-	for mi.Valid() {
-		keys = append(keys, string(splitIterUserKey(t, mi.Item().Entry().Key)))
-		mi.Next()
-	}
-	require.Equal(t, []string{"a", "b", "c"}, keys)
-
-	revLeft := &sliceIterator{entries: left.entries, reverse: true}
-	revRight := &sliceIterator{entries: right.entries, reverse: true}
-	rev := NewMergeIterator([]index.Iterator{revLeft, revRight}, true)
-	keys = keys[:0]
-	rev.Rewind()
-	for rev.Valid() {
-		keys = append(keys, string(splitIterUserKey(t, rev.Item().Entry().Key)))
-		rev.Next()
-	}
-	require.Equal(t, []string{"c", "b", "a"}, keys)
-}
-
-func TestMergeIteratorSeekAndClose(t *testing.T) {
-	left := &sliceIterator{entries: []*kv.Entry{
-		{Key: kv.InternalKey(kv.CFDefault, []byte("a"), 1)},
-		{Key: kv.InternalKey(kv.CFDefault, []byte("b"), 1)},
-		{Key: kv.InternalKey(kv.CFDefault, []byte("d"), 1)},
-	}}
-	right := &sliceIterator{entries: []*kv.Entry{
-		{Key: kv.InternalKey(kv.CFDefault, []byte("c"), 1)},
-	}}
-
-	mi := NewMergeIterator([]index.Iterator{left, right}, false)
-	mi.Seek(kv.InternalKey(kv.CFDefault, []byte("c"), 1))
-	require.True(t, mi.Valid())
-	require.Equal(t, "c", string(splitIterUserKey(t, mi.Item().Entry().Key)))
-	require.NoError(t, mi.Close())
-}
-
-func TestMergeIteratorEmptyInputIsSafe(t *testing.T) {
-	mi := NewMergeIterator(nil, false)
-	require.NotNil(t, mi)
-	require.NotPanics(t, func() { mi.Rewind() })
-	require.NotPanics(t, func() { mi.Seek([]byte("k")) })
-	require.NotPanics(t, func() { mi.Next() })
-	require.False(t, mi.Valid())
-	require.Nil(t, mi.Item())
-	require.NoError(t, mi.Close())
 }
 
 func TestLSMNewIterators(t *testing.T) {
@@ -176,52 +50,6 @@ func TestLSMNewIterators(t *testing.T) {
 		it.Seek(entry.Key)
 		_ = it.Close()
 	}
-
-}
-
-func TestConcatIteratorSeekAndNext(t *testing.T) {
-	dir := t.TempDir()
-	opt := &Options{
-		WorkDir:            dir,
-		MemTableSize:       1 << 20,
-		SSTableMaxSz:       1 << 20,
-		BlockSize:          4 << 10,
-		BloomFalsePositive: 0.01,
-	}
-	lsm := buildTestLSM(t, opt)
-	defer func() { require.NoError(t, lsm.Close()) }()
-
-	builder := table.NewBuilder(tableOptionsFor(opt))
-	builder.AddKey(kv.NewEntry(kv.InternalKey(kv.CFDefault, []byte("b"), 1), []byte("vb")))
-	builder.AddKey(kv.NewEntry(kv.InternalKey(kv.CFDefault, []byte("d"), 1), []byte("vd")))
-	tbl, err := table.Open(lsm.levels, vfs.FileNameSSTable(dir, 100), builder)
-	require.NoError(t, err)
-	require.NotNil(t, tbl)
-	defer func() { _ = tbl.DecrRef() }()
-
-	ci := NewConcatIterator([]*table.Table{tbl}, &index.Options{IsAsc: true})
-
-	ci.Rewind()
-	if !ci.Valid() {
-		t.Fatalf("expected concat iterator to be valid after rewind")
-	}
-	if ci.Item() == nil {
-		t.Fatalf("expected non-nil item")
-	}
-
-	ci.Seek(kv.InternalKey(kv.CFDefault, []byte("c"), 1))
-	if !ci.Valid() {
-		t.Fatalf("expected concat iterator valid after seek")
-	}
-	got := splitIterUserKey(t, ci.Item().Entry().Key)
-	if string(got) != "d" {
-		t.Fatalf("expected seek to land on d, got %q", string(got))
-	}
-
-	ci.Next()
-	if ci.Valid() {
-		t.Fatalf("expected iterator to be exhausted")
-	}
 }
 
 func TestBlockIteratorReverse(t *testing.T) {
@@ -233,7 +61,6 @@ func TestBlockIteratorReverse(t *testing.T) {
 	builderOpt.BlockSize = 64
 	builder := table.NewBuilder(tableOptionsFor(&builderOpt))
 
-	// Add test data
 	for i := range 10 {
 		key := kv.InternalKey(kv.CFDefault, []byte{byte('a' + i)}, 1)
 		value := []byte{byte('v'), byte('0' + i)}
@@ -246,7 +73,6 @@ func TestBlockIteratorReverse(t *testing.T) {
 	require.NotNil(t, tbl)
 	defer func() { _ = tbl.DecrRef() }()
 
-	// Test forward iteration
 	forwardIter := tbl.NewIterator(&index.Options{IsAsc: true})
 	defer func() { require.NoError(t, forwardIter.Close()) }()
 
@@ -259,7 +85,6 @@ func TestBlockIteratorReverse(t *testing.T) {
 	}
 	require.Equal(t, "abcdefghij", string(forwardKeys))
 
-	// Test reverse iteration
 	reverseIter := tbl.NewIterator(&index.Options{IsAsc: false})
 	defer func() { require.NoError(t, reverseIter.Close()) }()
 
@@ -282,7 +107,6 @@ func TestTableIteratorReverseSeek(t *testing.T) {
 	builderOpt.BlockSize = 64
 	builder := table.NewBuilder(tableOptionsFor(&builderOpt))
 
-	// Add test data with multiple blocks
 	for i := range 20 {
 		key := kv.InternalKey(kv.CFDefault, []byte{byte('a' + i)}, 1)
 		value := bytes.Repeat([]byte{byte('v'), byte('0' + i%10)}, 24)
@@ -295,7 +119,6 @@ func TestTableIteratorReverseSeek(t *testing.T) {
 	require.NotNil(t, tbl)
 	defer func() { _ = tbl.DecrRef() }()
 
-	// Test forward seek
 	forwardIter := tbl.NewIterator(&index.Options{IsAsc: true})
 	defer func() { require.NoError(t, forwardIter.Close()) }()
 
@@ -304,7 +127,6 @@ func TestTableIteratorReverseSeek(t *testing.T) {
 	e := forwardIter.Item().Entry()
 	require.Equal(t, byte('e'), splitIterUserKey(t, e.Key)[0])
 
-	// Test reverse seek
 	reverseIter := tbl.NewIterator(&index.Options{IsAsc: false})
 	defer func() { require.NoError(t, reverseIter.Close()) }()
 
@@ -313,7 +135,6 @@ func TestTableIteratorReverseSeek(t *testing.T) {
 	e = reverseIter.Item().Entry()
 	require.Equal(t, byte('e'), splitIterUserKey(t, e.Key)[0])
 
-	// Continue reverse iteration
 	var keys []byte
 	for i := 0; i < 5 && reverseIter.Valid(); i++ {
 		e := reverseIter.Item().Entry()
@@ -332,7 +153,6 @@ func TestTableIteratorReverseMultiBlock(t *testing.T) {
 	builderOpt.BlockSize = 64
 	builder := table.NewBuilder(tableOptionsFor(&builderOpt))
 
-	// Add enough data to create multiple blocks
 	for i := range 30 {
 		key := kv.InternalKey(kv.CFDefault, []byte{byte('a' + i)}, 1)
 		value := bytes.Repeat([]byte{byte('v')}, 48)
@@ -345,7 +165,6 @@ func TestTableIteratorReverseMultiBlock(t *testing.T) {
 	require.NotNil(t, tbl)
 	defer func() { _ = tbl.DecrRef() }()
 
-	// Test reverse iteration across multiple blocks
 	reverseIter := tbl.NewIterator(&index.Options{IsAsc: false})
 	defer func() { require.NoError(t, reverseIter.Close()) }()
 
@@ -361,7 +180,6 @@ func TestTableIteratorReverseMultiBlock(t *testing.T) {
 		count++
 	}
 
-	// Verify we got all keys in reverse order
 	require.Equal(t, 30, len(keys))
 	for i := range 30 {
 		expected := byte('a' + 29 - i)
